Add -Wsign-conversion to build flags and fix all violations
Enables -Wsign-conversion -Werror across the library, tests, and
benchmarks. Fixes all violations in rmm-owned code:

- fixed_size_memory_resource.hpp: block_gen lambda index int -> size_t
- statistics_resource_adaptor.hpp: cast bytes to int64_t for counter +=/-=
- mr_ref_test.hpp: fix index type and std::next difference_type cast
- mr_ref_test_mt.hpp: cast num_devices to size_t for vector::reserve
- tracking_mr_tests.cpp: cast loop index for vector subscript
- prefetch_tests.cpp: cast size_t to ptrdiff_t for std::next
- log_parser.hpp: cast stoll result to uintptr_t

rapidcsv.h and logger_tests.cpp (which includes it via log_parser.hpp)
are suppressed per-target as rapidcsv.h is vendored third-party code.

+++ cpp/include/rmm/mr/statistics_resource_adaptor.hpp
@@ -227,7 +227,7 @@ class statistics_resource_adaptor final : public device_memory_resource {
       write_lock_t lock(mtx_);
 
       // Increment the allocation_count_ while we have the lock
-      counter_stack_.top().first += bytes;
+      counter_stack_.top().first += static_cast<int64_t>(bytes);
       counter_stack_.top().second += 1;
     }
 
@@ -249,7 +249,7 @@ class statistics_resource_adaptor final : public device_memory_resource {
       write_lock_t lock(mtx_);
 
       // Decrement the current allocated counts.
-      counter_stack_.top().first -= bytes;
+      counter_stack_.top().first -= static_cast<int64_t>(bytes);
       counter_stack_.top().second -= 1;
     }
   }
